Remove my-github.js from "pull-requests" and "others" pages
This code will try to use an element for the repos which only
exists in the my-github.html page, so it will break in the other
pages which confuses the students. Part of the instructions for
the pull-requests.html page is to add a link to the pull-requests.js
file, so there's no need for the my-github.js file to be included.